requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.0.9
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==21.11.1
Automat==20.2.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
cssselect==1.1.0
daphne==3.0.2
Deprecated==1.2.13
Django==3.1.14
django-ckeditor==6.2.0
django-cors-headers==3.11.0
django-filter==21.1
django-js-asset==1.2.2
django-multi-captcha-admin==2.0.0
django-ranged-response==0.2.0
django-redis==5.2.0
django-simple-captcha==0.5.14
django-simpleui==2022.1
djangorestframework==3.12.4
djangorestframework-csv==2.1.1
django-silk==4.2.0
djangorestframework-simplejwt==5.0.0
dnspython==2.2.0
drf-yasg==1.20.0
eventlet==0.33.0
frozenlist==1.3.0
geoip2==4.5.0
greenlet==1.1.2
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.3
lxml==4.7.1
MarkupSafe==2.0.1
maxminddb==2.2.0
multidict==6.0.2
mysqlclient==2.1.0
packaging==21.3
paypalrestsdk==1.13.1
Pillow==9.0.0
prompt-toolkit==3.0.24
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodomex==3.9.4
PyJWT==2.3.0
pyOpenSSL==19.1.0
pyparsing==3.0.7
pyquery==1.4.3
pysnowflake==0.1.3
python-alipay-sdk==3.0.4
pytz==2021.3
qrcode==7.3.1
redis==4.1.1
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tornado==6.1
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing_extensions==4.0.1
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
wechatpayv3==1.2.12
wrapt==1.13.3
yarl==1.7.2
zope.interface==5.4.0
